+++ src/whatsapp/controllers/chatwoot.controller.ts
@@ -94,4 +94,10 @@ export class ChatwootController {
 
     return chatwootService.receiveWebhook(instance, data);
   }
+
+  public async newInstance(data: any) {
+    const chatwootService = new ChatwootService(waMonitor, this.configService);
+
+    return chatwootService.newInstance(data);
+  }
 }