+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed issue where it was not possible to open a conversation when sent at first by me on my cell phone in chatwoot
 * Now it only updates the contact name if it is the same as the phone number in chatwoot
 * Now accepts all chatwoot inbox templates
+* Command to create new instances set to /new_instance:<NAME>:<NUMBER>
 
 ### Integrations